Problem: reliance on a generic allocator
Generic allocator like malloc/free pair is not aware what it is
allocating for.

However, there are scenarios where it is desirable for an allocator to
be aware and allocate appropriately (for example, in different pools of
memory)

Solution: introduce intention/context into allocators
by providing high-level ALLOC_FN and ALLOC_FREE_FN API.

It defaults to wrappers for MALLOC_FN and FREE_FN but allows one to
override to gain visibility into what's going on.